Guard Ultimate Sort of Working
The Pillow Talk Ult for the Guard works but not in all the ways I want it to. I want some way to make sure it ALWAYS spawns in front of the player and not all enemies can hurt the pillow. Wassup with dat broski??????

Oh also I commented a fuckton of my code earlier and that's all in this push too. ye