Guard's Ultimate Ability got some tweaks
Okay so the Guard's Ult redirects damage but not the way I'd like so I would like to figure out a better way. It sees if damage is dealt, undoes it, then does that damage to the pillow. 

I've also added pushback every second for the duration of the pillow. This has the fucking bizarre glitch of enemies not going back on their path when the pushback is done? The fuck is up with that my dude? 

My commit messages have gotten slightly more informative but I need to make sure I still keep them vulgar and fucking stupid sometimes so that nobody thinks I'm a professional or something.